Made changes after the rewiew
