+++ project/thread_pool.py
@@ -1,6 +1,4 @@
 import threading
-import concurrent.futures
-import itertools
 from typing import List, Callable, Tuple
 
 
@@ -61,27 +59,3 @@ class ThreadPool:
         # Wait for all threads to complete
         for thread in self.threads:
             thread.join()
-
-
-def sum_cartesian_product(sets: List[List[int]]) -> int:
-    """
-    Computes the sum of the full Cartesian product of integer sets in parallel.
-
-    :param sets: A list of integer sets for the Cartesian product.
-    :return: The sum of all elements in the Cartesian product.
-    """
-    # Generate the full Cartesian product
-    cartesian_product = itertools.product(*sets)
-
-    # Function to sum a combination of integers
-    def sum_combination(combination: Tuple[int]) -> int:
-        return sum(combination)
-
-    total_sum = 0
-    # Parallel summation
-    with concurrent.futures.ThreadPoolExecutor() as executor:
-        results = list(executor.map(sum_combination, cartesian_product))
-
-    # Compute the total sum
-    total_sum = sum(results)
-    return total_sum
